+++ src/lib/template/_article.svelte
@@ -3,10 +3,10 @@
   /* eslint-disable @typescript-eslint/ban-ts-comment */
   // @ts-nocheck
   import Post from '$lib/layouts/post.svelte';
-  import ImgZ from '$lib/components/image_zoom.svelte';
+  import ImgZoom from '$lib/components/image_zoom.svelte';
   import Video from '$lib/components/video.svelte';
-  import CodeCopy from '$lib/components/code_copybtn.svelte';
-  import InfoBox from '$lib/components/infobox.svelte';
+  import CodeCopy from '$lib/components/code_copy.svelte';
+  import InfoBox from '$lib/components/info_box.svelte';
 
   /*<!-- :QWER IMPORTS: -->*/
 </script>